GraphSignals

GraphSignals.FeaturedGraph

GraphSignals.NullGraph

GraphSignals.SparseGraph

GraphSignals._to_csc_index

GraphSignals.adjacency_list

GraphSignals.adjacency_list

GraphSignals.adjacency_list

GraphSignals.aggregate_index

GraphSignals.edge_feature

GraphSignals.edge_feature

GraphSignals.edge_scatter

GraphSignals.global_feature

GraphSignals.global_feature

GraphSignals.graph

GraphSignals.graph

GraphSignals.graph

GraphSignals.has_edge_feature

GraphSignals.has_edge_feature

GraphSignals.has_global_feature

GraphSignals.has_global_feature

GraphSignals.has_graph

GraphSignals.has_graph

GraphSignals.has_node_feature

GraphSignals.has_node_feature

GraphSignals.incident_edges

GraphSignals.mask

GraphSignals.neighbor_scatter

GraphSignals.node_feature

GraphSignals.node_feature

GraphSignals.order_edges

LightGraphs.ne

LightGraphs.ne

LightGraphs.neighbors

LightGraphs.nv

LightGraphs.nv

GraphSignals.FeaturedGraph — Type

FeaturedGraph(g, [mt]; nf, ef, gf, directed)

A type representing a graph structure and storing also arrays that contain features associated to nodes, edges, and the whole graph.

A FeaturedGraph can be constructed out of different objects g representing the connections inside the graph. When constructed from another featured graph fg, the internal graph representation is preserved and shared.

Arguments

g: Data representing the graph topology. Possible type are

An adjacency matrix.

An adjacency list.

A LightGraphs' graph, i.e. SimpleGraph, SimpleDiGraph from LightGraphs, or SimpleWeightedGraph, SimpleWeightedDiGraph from SimpleWeightedGraphs.

An AbstractFeaturedGraph object.

mt: matrix type for g in matrix form. if graph is in matrix form, mt is recorded as one of :adjm, :laplacian, :normalized or :scaled.

nf: Node features.

ef: Edge features.

gf: Global features.

Usage

using GraphSignals, CUDA

# Construct from adjacency list representation
g = [[2,3], [1,4,5], [1], [2,5], [2,4]]
fg = FeaturedGraph(g)

# Number of nodes and edges
nv(fg)  # 5
ne(fg)  # 10

# From a LightGraphs' graph
fg = FeaturedGraph(erdos_renyi(100, 20))

# Copy featured graph while also adding node features
fg = FeaturedGraph(fg, nf=rand(100, 5))

# Send to gpu
fg = fg |> cu

See also graph, node_feature, edge_feature, and global_feature

GraphSignals.NullGraph — Type

NullGraph()

Null object for FeaturedGraph.

GraphSignals.SparseGraph — Type

SparseGraph(A, directed)

A sparse graph structure represents by sparse matrix. A directed graph is represented by a sparse matrix, of which column index as source node index and row index as sink node index.

GraphSignals._to_csc_index — Method

Transform a regular cartesian index A[i, j] into a CSC-compatible index spA.nzval[idx].

GraphSignals.adjacency_list — Method

adjacency_list(::AbstractFeaturedGraph)

Get adjacency list of graph.

GraphSignals.adjacency_list — Method

adjacency_list(adj)

Transform a adjacency matrix into a adjacency list.

GraphSignals.aggregate_index — Function

aggregate_index(sg; direction=:undirected, kind=:edge)

Generate index structure for scatter operation.

Arguments

sg::SparseGraph: The reference graph.

direction::Symbol: The direction of an edge to be choose to aggregate. It must be one of :inward and :outward.

kind::Symbol: To aggregate feature upon edge or vertex. It must be one of :edge and :vertex.

GraphSignals.edge_feature — Method

edge_feature(::AbstractFeaturedGraph)

Get edge feature attached to graph.

GraphSignals.edge_scatter — Method

edge_scatter(aggr, E, sg, direction=:undirected)

Scatter operation for aggregating edge feature into vertex feature.

Arguments

aggr: aggregating operators, e.g. +.

E: Edge features with size of (#feature, #edge).

sg::SparseGraph: The reference graph.

direction::Symbol: The direction of an edge to be choose to aggregate. It must be one of :undirected, :inward and :outward.

GraphSignals.global_feature — Method

global_feature(::AbstractFeaturedGraph)

Get global feature attached to graph.

GraphSignals.graph — Method

Support adjacency matrix

GraphSignals.graph — Method

graph(::AbstractFeaturedGraph)

Get referenced graph.

GraphSignals.has_edge_feature — Method

has_edge_feature(::AbstractFeaturedGraph)

Check if edge feature is available or not.

GraphSignals.has_global_feature — Method

has_global_feature(::AbstractFeaturedGraph)

Check if global feature is available or not.

GraphSignals.has_graph — Method

has_graph(::AbstractFeaturedGraph)

Check if graph is available or not.

GraphSignals.has_node_feature — Method

has_node_feature(::AbstractFeaturedGraph)

Check if node feature is available or not.

GraphSignals.incident_edges — Method

incident_edges(sg, i)

Return the edges incident to vertex i in sparse graph sg.

Arguments

sg::SparseGraph: sparse graph to query.

i: vertex index.

GraphSignals.mask — Method

mask(fg, m)

A syntax sugar for masking graph.

Returns a GraphMask.

GraphSignals.neighbor_scatter — Method

neighbor_scatter(aggr, X, sg, direction=:undirected)

Scatter operation for aggregating neighbor vertex feature together.

Arguments

aggr: aggregating operators, e.g. +.

X: Vertex features with size of (#feature, #vertex).

sg::SparseGraph: The reference graph.

direction::Symbol: The direction of an edge to be choose to aggregate. It must be one of :undirected, :inward and :outward.

GraphSignals.node_feature — Method

node_feature(::AbstractFeaturedGraph)

Get node feature attached to graph.

GraphSignals.order_edges — Method

Order the edges in a graph by giving a unique integer to each edge.

LightGraphs.ne — Method

ne(::AbstractFeaturedGraph)

Get edge number of graph.

LightGraphs.neighbors — Method

neighbors(sg, i)

Return the neighbors of vertex i in sparse graph sg.

Arguments

sg::SparseGraph: sparse graph to query.

i: vertex index.

LightGraphs.nv — Method

nv(::AbstractFeaturedGraph)

Get node number of graph.